requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
python-dotenv==1.0.1
langchain-openai==0.1.1
langchain-community==0.0.38
duckduckgo-search==5.3.0